requirements.txt
aqt[qt6]
openai
python-dotenv
